add ID stage
read register at the second pipeline stage.
It shows that the pipeline simulation works. But now, the
function calling order determines the circuit's behavior.

For example, the sequental logic simulation function should be
called in front of the combinational logc, so that the dffs works
at the rising edge and the combinational logic works after that.

In exu(), the rf() is called before decode().

Next, I should change the clk to the following form,

--__--__--

Each high and low has two units. In this desgin, for now, all the
sequential logic components should work at the rising edge and the
combinational logic compents work at the next unit in a high level.